Add exclamation point to duplicate error message text.

+++ sbbs3/chk_ar.cpp
@@ -2,7 +2,7 @@
 
 /* Synchronet ARS checking routine */
 
-/* $Id: chk_ar.cpp,v 1.12 2004/10/17 06:46:31 rswindell Exp $ */
+/* $Id: chk_ar.cpp,v 1.13 2004/12/01 03:06:40 rswindell Exp $ */
 
 /****************************************************************************
  * @format.tab-size 4		(Plain Text/Source Code File Header)			*
@@ -394,7 +394,7 @@ bool sbbs_t::ar_exp(uchar **ptrptr, user_t* user)
 				break;
 			case AR_PCR:	/* post/call ratio (by percentage) */
 				if(user->logons>user->posts
-					&& (!user->posts || (100/(user->logons/user->posts))<(long)n))
+					&& (!user->posts || (100/((float)user->logons/user->posts))<(long)n))
 					result=_not;
 				else
 					result=!_not;
@@ -406,7 +406,7 @@ bool sbbs_t::ar_exp(uchar **ptrptr, user_t* user)
 				l=user->dlb;
 				if(!l) l=1;
 				if(user->dlb>user->ulb
-					&& (!user->ulb || (100/(l/user->ulb))<n))
+					&& (!user->ulb || (100/((float)l/user->ulb))<n))
 					result=_not;
 				else
 					result=!_not;
@@ -418,7 +418,7 @@ bool sbbs_t::ar_exp(uchar **ptrptr, user_t* user)
 				i=user->dls;
 				if(!i) i=1;
 				if(user->dls>user->uls
-					&& (!user->uls || (100/(i/user->uls))<n))
+					&& (!user->uls || (100/((float)i/user->uls))<n))
 					result=_not;
 				else
 					result=!_not;
